Changed decimal types to double.

+++ src/ElevateOTT.Application/Features/Content/Videos/Commands/UpdateVideo/UpdateVideoCommand.cs
@@ -13,10 +13,10 @@ public class UpdateVideoCommand : BaseAssetDto, IRequest<Envelope<string>>
     public bool Mp4Support { get; set; }
 
     public bool HasOneTimePurchasePrice { get; set; }
-    public decimal OneTimePurchasePrice { get; set; }
+    public double OneTimePurchasePrice { get; set; }
     public bool HasRentalPrice { get; set; }
     public RentalDuration RentalDuration { get; set; }
-    public decimal RentalPrice { get; set; }
+    public double RentalPrice { get; set; }
 
     public Guid? TrailerVideoId { get; set; }
     public Guid? FeaturedCategoryVideoId { get; set; }
